Initial implementation of sidetone logic. Unit tests for modargs parsing.
The logic does not work properly due to a bug/feature with ALSA mixers
(outdated values are read from mixer elements due to the fact that we're using
a separate mixer handle than the ALSA sink). For testing purposes, the mixer
handle could be re-opened always before requesting volumes from elements. If
this (ugly hack) is done, the logic seems to work properly (but also seems to
cause an assert eventually if pa_alsa_old_open_mixer is used). Using only one
handle for accessing the mixer should fix the problem. Thus, it might be
necessary to add facilities for requesting the mixer handle from the ALSA sink.

Howling control probably still needs to be used somehow.