music & voice: Forward the sink rewind request to the master sink only if the rewind request is more than 0 bytes.
Passing 0 bytes to the sink input that connects the virtual
sink to the master sink isn't a good idea, because the sink
input interprets requests of 0 bytes so that the maximum
amount should be rewound. This is not what was actually
requested, and doing this can cause gaps in the audio.