mainvolume: Disable volume proxy code.
This makes the module useless, because the module won't anymore have
any access to the actual volumes. The reason for this patch is that
I'd like to drop the volume proxy from pulseaudio (it's not in
upstream, and it's not going to be either, at least in the current
form).

It may turn out that in the short term the volume proxy and other
related hacks in pulseaudio are the only viable solution to make
volume handling work well in Nemo, but for now I just want to make
this module compile without porting all the necessary hacks to
pulseaudio.