New module: module-stream-restore-nemo.
It's a copy of Harmattan's module-stream-restore. In Harmattan, the
module was shipped in the main pulseaudio package. It contained major
patches, adding "route volumes" and integrating with the "volume
proxy". Those patches are not suitable for upstreaming, so instead of
patching the main pulseaudio package, I copied the Harmattan version
here and renamed it to module-stream-restore-nemo.

The route volume functionality achieves separate volumes for music and
call streams for each output route. That's useful in Nemo, and it
would be good to have the feature in upstream also, but this is
probably not the best way to achieve it.

The volume proxy makes it possible for the mainvolume module to
interact with the stream-restore module.