voice: Don't return silence from hw-sink-input if the sink asks for an "odd" amount of data.
The original logic was added in commit
61fb3f326d7ac1b1d2b913af717d55965822df1c, and refined in
5b0b28268eb936eea09987dd246d6ce096ebd5ec. That logic supposedly helped
with AEC issues on N9/Harmattan. On N900/Nemo, however, it completely
breaks the downlink audio. It remains to be seen if the AEC issue
resurfaces...

I think the real fix would be to force the alsa sink to never request
anything else than the fragment size (or a multiple of it). Adding
support for that would be non-trivial, but far from impossible, I
believe.