[mainvolume] Update high volume and listening time notifier.
Every mode change triggers NotifyHighVolume with value 0. When proper
volume for new mode is known, NotifyHighVolume may be signalled again,
if new mode hash high-volume-step defined and restored volume is same
or higher than high-volume-step.

Listening time notifier configuration is changed so that sink-list
values are matched as starts-with. So bluez_sink. matches to
bluez_sink.foo etc.

For both sink-list and mode-list delimiter is now ";".